Change comparison procedures to understand extended number types.
This is a pretty big and very ugly change, for a few reasons.

First, we make comparisons between fixnum and flonums more carefully:
We might be on a 64-bit system; simply coercing the fixnum to a flonum
would case 10 bits of data loss there.  Instead we coerce the fixnum
to a bignum, do the same to the integer part of the flonum and compare
the two.  If they're equal, we simply check the sign and the
fractional part to break the tie.

To compare ratnums with flonums we multiply the numerator and the
flonum by the denominator, and we shift both by the number of bits in
the flonum fraction.  Afterwards, we simply compare the resulting
bignums.

Finally, in order to keep things manageable we have only one
comparator function (basic_cmp) that does the hard work.  The various
comparison primitives are expressed in terms of basic_cmp.